+++ tests/Fixtures/Node/NodeWithHandleAttribute.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Duon\Cms\Tests\Fixtures\Node;
 
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Handle;
 
 #[Handle('node-with-custom-handle-attribute')]
-class NodeWithHandleAttribute implements HasTitle
+class NodeWithHandleAttribute implements Title
 {
 	public function title(): string
 	{