+++ tests/Fixtures/Node/NodeWithNameAttribute.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Duon\Cms\Tests\Fixtures\Node;
 
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Label;
 
 #[Label('Node With Custom Name Attribute')]
-class NodeWithNameAttribute implements HasTitle
+class NodeWithNameAttribute implements Title
 {
 	public function title(): string
 	{