+++ tests/Fixtures/Node/NodeWithPermissionAttribute.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Duon\Cms\Tests\Fixtures\Node;
 
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Permission;
 
 #[Permission([
 	'read' => 'me',
 ])]
-class NodeWithPermissionAttribute implements HasTitle
+class NodeWithPermissionAttribute implements Title
 {
 	public function title(): string
 	{