+++ tests/Fixtures/Node/NodeWithRenderAttribute.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Duon\Cms\Tests\Fixtures\Node;
 
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Render;
 
 #[Render('template-defined-by-render-attribute')]
-class NodeWithRenderAttribute implements HasTitle
+class NodeWithRenderAttribute implements Title
 {
 	public function title(): string
 	{