+++ tests/Fixtures/Node/TestBlock.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Duon\Cms\Tests\Fixtures\Node;
 
 use Duon\Cms\Field\Text;
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Label;
 use Duon\Cms\Schema\Translate;
 
 #[Label('Test Block')]
-class TestBlock implements HasTitle
+class TestBlock implements Title
 {
 	#[Label('Title')]
 	#[Translate]