+++ tests/Fixtures/Node/TestMediaDocument.php
@@ -8,7 +8,7 @@ use Duon\Cms\Field\Grid;
 use Duon\Cms\Field\Image;
 use Duon\Cms\Field\Option;
 use Duon\Cms\Field\Video;
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Columns;
 use Duon\Cms\Schema\Label;
 use Duon\Cms\Schema\Multiple;
@@ -17,7 +17,7 @@ use Duon\Cms\Schema\Translate;
 use Duon\Cms\Schema\TranslateFile;
 
 #[Label('Test Media Document')]
-class TestMediaDocument implements HasTitle
+class TestMediaDocument implements Title
 {
 	#[Label('Gallery')]
 	#[Multiple]