+++ tests/Fixtures/Node/TestNodeWithMatrix.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 namespace Duon\Cms\Tests\Fixtures\Node;
 
 use Duon\Cms\Field\Text;
-use Duon\Cms\Node\Contract\HasTitle;
+use Duon\Cms\Node\Contract\Title;
 use Duon\Cms\Schema\Label;
 use Duon\Cms\Schema\Required;
 use Duon\Cms\Schema\Translate;
 
 #[Label('Test Node With Matrix')]
-class TestNodeWithMatrix implements HasTitle
+class TestNodeWithMatrix implements Title
 {
 	#[Label('Titel'), Required, Translate]
 	protected Text $title;